Display unread log entry count
Using a DataTrigger and custom value converter to display the number of
unread log entries alongside the log subscription name (and bold the
name if count > 0).
Added INotifyPropertyChanged implementations and custom
ObservableCollection<T> to ensure UI is updated as log entries are added
and then marked as read.
Moved MarkAsReadAndSave work to a background thread - still some work to
do here.